Use long long for all helper methods

+++ src/pyjion/intrins.h
@@ -314,9 +314,9 @@ int PyJit_SetupAnnotations(PyFrameObject* frame);
 
 PyObject* PyJit_GetListItemReversed(PyObject* list, size_t index);
 
-double PyJit_LongTrueDivide(long x, long y);
-long PyJit_LongFloorDivide(long x, long y);
-long PyJit_LongMod(long x, long y);
-long PyJit_LongPow(long x, long y);
+double PyJit_LongTrueDivide(long long x, long long y);
+long long PyJit_LongFloorDivide(long long x, long long y);
+long long PyJit_LongMod(long long x, long long y);
+long long PyJit_LongPow(long long x, long long y);
 
 #endif
